Major upstream changes:
- perf: project usage queries with column-set struct, switch primary keys to int64 across entities/migrations
- fix: correct Claude cache rate formula (introduces calculateCacheRate util)
- Remove model price remote sync service and analysis API
- Remove credential stats card from upstream UI

Fork-specific decisions:
- Keep our restored credential stats endpoints and UI (internal/api/usage_credentials.go, usage_source_resolution.go, web/.../CredentialStatsCard.*).
- Keep getAnalysisApiStats/calculateAnalysisModelCost utils and UsageAnalysis* types — still used by UsagePage analysis path.
- Adapt uintToString helper to int64 after upstream changed IDs.
- Force github.com/glebarez/sqlite in new int64-primary-keys migration test instead of upstream's gorm.io/driver/sqlite.
- Backfill missing zh-TW translations for load_more / model_price_sync keys.
- Restore .github/workflows to ci.yml + release.yml only; drop upstream's binary/docker publish workflows.

Verified: go build, go test ./cmd/... ./internal/..., npm test, cross-compile linux/amd64+linux/arm64.

+++ internal/entities/usage_event.go
@@ -4,7 +4,7 @@ import "time"
 
 // UsageEvent 是落库后的单条 usage 请求事件实体。
 type UsageEvent struct {
-	ID              uint      `gorm:"primaryKey;index:idx_usage_events_timestamp_id,sort:desc,priority:2;index:idx_usage_events_auth_type_auth_index_id,priority:3;index:idx_usage_events_auth_type_source_id,priority:3"`
+	ID              int64     `gorm:"primaryKey;index:idx_usage_events_timestamp_id,sort:desc,priority:2;index:idx_usage_events_auth_type_auth_index_id,priority:3;index:idx_usage_events_auth_type_source_id,priority:3"`
 	EventKey        string    `gorm:"uniqueIndex:uniq_usage_events_event_key"`
 	APIGroupKey     string    `gorm:"index:idx_usage_events_trim_api_group_key,expression:TRIM(api_group_key)"`
 	Provider        string    `gorm:"column:provider;index:idx_usage_events_trim_provider,expression:TRIM(provider)"`
